requirements.txt
Django==2.2.5
django-rest-framework==0.1.0
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
django-debug-toolbar
